vending_machine-main/vending_machine-main/Mysql/requirements.txt
# requirements.txt
pymysql==1.1.0
python-dotenv==1.0.0